Mon 10989 - CRSF token vulnerability (#10150)
* fix(secu) add fcts relative to csrf token validation

* fix(secu) check csrf token validity when performing action in list pages
(contact, contactTemplate, contactGroup, timeperiod)

* fix(secu) check csrf token validity when performing action in list pages
(host, hostGroup, hostTemplate, hostCategories)

* fix(secu) check csrf token validity when performing action in list pages
(serviceByHost, serviceByHostGroup, serviceTemplate, serviceGroup, serviceCategories, metaservice)

* fix(secu) check csrf token validity when performing action in list pages
(command, connector)

* fix(secu) check csrf token validity when performing action in list pages
(escalation, hostDependency, hostGroupDependency, serviceDependency,
serviceGroupDependency, metaserviceDependency)

* fix(secu) check csrf token validity when performing action in list pages
(snmp traps, groups, manufacturer)

* fix(secu) check csrf token validity when performing action in list pages
(conf/pollers > centreonBroker, nagios, reources, servers)

* fix(secu) check csrf token validity when performing action in list pages
(monitoring/performances > componentTemplates, graphTemplates,
virtualMetrics)

* fix(secu) check csrf token validity when performing action in list pages
(monitoring/dowtimes > comments/commentHost, dowtime, recurrentDowtime)

* fix(secu) check csrf token validity when performing action in list pages
(admin/parameters > ldap, images)

* fix(secu) check csrf token validity when performing action in list pages
(admin/acl > actionAccess, groupAccess, menuAccess, resourcesAccess)

* fix(secu) add token to individual enable/disable action icon

* fix(secu) corrections following review comments

* fix(secu) corrections following review comments

* Update www/lib/HTML/QuickForm/HTML_QuickFormCustom.php

* fix(secu) modify token verification and deletion following code review

+++ www/include/configuration/configObject/traps/traps.php
@@ -1,4 +1,5 @@
 <?php
+
 /*
  * Copyright 2005-2015 Centreon
  * Centreon is developped by : Julien Mathis and Romain Le Merlus under
@@ -95,14 +96,26 @@ switch ($o) {
         }
         break;
     case TRAP_DUPLICATE:
-        if (!in_array(false, $selectIds) && !in_array(false, $duplicateNbr)) {
-            $trapObj->duplicate($selectIds, $duplicateNbr);
+        purgeOutdatedCSRFTokens();
+        if (isCSRFTokenValid()) {
+            purgeCSRFToken();
+            if (!in_array(false, $selectIds) && !in_array(false, $duplicateNbr)) {
+                $trapObj->duplicate($selectIds, $duplicateNbr);
+            }
+        } else {
+            unvalidFormMessage();
         }
         require_once($path . "listTraps.php");
         break;
     case TRAP_DELETE:
-        if (!in_array(false, $selectIds)) {
-            $trapObj->delete($selectIds);
+        purgeOutdatedCSRFTokens();
+        if (isCSRFTokenValid()) {
+            purgeCSRFToken();
+            if (!in_array(false, $selectIds)) {
+                $trapObj->delete($selectIds);
+            }
+        } else {
+            unvalidFormMessage();
         }
         require_once($path . "listTraps.php");
         break; #
